requirements.txt
langchain
langchain-core
langchain-community
langgraph
python-dotenv
SQLAlchemy
fastapi
uvicorn
langchain-groq
psycopg2-binary
